I have updated & improved multiple scripts.  I have not been well & I am still not so I've not kept as much track of my changes.  LOL but that is one of the wonderful things about source control & especially git.  I do not really have to :)

+++ scripts/my-cli-canvas.tcsh
@@ -34,9 +34,9 @@ sssss
 			--tab-with-profile="projects" --title="console" --working-directory="/projects/console" \
 			--tab-with-profile="projects" --title="gtk" --working-directory="/projects/gtk" \
 			--tab-with-profile="projects" --title="GTK-PHP-IDE" --working-directory="/projects/gtk/GTK-PHP-IDE/GTK-PHP-IDE" \
-			--tab-with-profile="projects" --title="Alacast" --working-directory="/projects/www/Alacast/version2" \
+			--tab-with-profile="projects" --title="Alacast" --working-directory="/projects/gtk/Alacast" \
 			--tab-with-profile="projects" --title="www" --working-directory="/projects/www" \
-			--tab-with-profile="projects" --title="uberChicGeekChick.Com" --working-directory="/projects/www/MyWebDesigns/uberChicGeekChick.Com" \
+			--tab-with-profile="projects" --title="uberChicGeekChick.Com" --working-directory="/projects/www/MyWebDesigns/mirrors/uberChicGeekChick.Com" \
 			--tab-with-profile="projects" --title="my media" --working-directory="/projects/media" \
 			--tab-with-profile="projects" --title="my games" --working-directory="/projects/games" \
 		${argv} &
@@ -69,11 +69,9 @@ tt		shift
 			--tab-with-profile="projects" --title="projects" --working-directory="/projects" \
 			--tab-with-profile="projects" --title="reference" --working-directory="/projects/reference" \
 			--tab-with-profile="projects" --title="srv" --working-directory="/srv" \
-			--tab-with-profile="projects" --title="GTK-PHP-IDE" --working-directory="/projects/gtk/GTK-PHP-IDE/GTK-PHP-IDE" \
 			--tab-with-profile="projects" --title="ssh" --working-directory="/projects/ssh" \
 			--tab-with-profile="projects" --title="www" --working-directory="/projects/www" \
-			--tab-with-profile="projects" --title="Alacast" --working-directory="/projects/www/Alacast/version2" \
-			--tab-with-profile="projects" --title="uberChicGeekChick.Com" --working-directory="/projects/www/MyWebDesigns/uberChicGeekChick.Com" \
+			--tab-with-profile="projects" --title="uberChicGeekChick.Com" --working-directory="/projects/www/MyWebDesigns/mirrors/uberChicGeekChick.Com" \
 		${argv} &
 		breaksw
 	
@@ -87,10 +85,9 @@ tt		shift
 			--tab-with-profile="projects" --title="projects" --working-directory="/projects" \
 			--tab-with-profile="projects" --title="ssh" --working-directory="/projects/ssh" \
 			--tab-with-profile="projects" --title="gtk" --working-directory="/projects/gtk" \
-			--tab-with-profile="projects" --title="GTK-PHP-IDE" --working-directory="/projects/gtk/GTK-PHP-IDE/GTK-PHP-IDE" \
 			--tab-with-profile="projects" --title="console" --working-directory="/projects/console" \
-			--tab-with-profile="projects" --title="Alacast" --working-directory="/projects/www/Alacast/version2" \
-			--tab-with-profile="projects" --title="uberChicGeekChick.Com" --working-directory="/projects/www/MyWebDesigns/uberChicGeekChick.Com" \
+			--tab-with-profile="projects" --title="Alacast" --working-directory="/projects/gtk/Alacast" \
+			--tab-with-profile="projects" --title="GTK-PHP-IDE" --working-directory="/projects/gtk/GTK-PHP-IDE/GTK-PHP-IDE" \
 		${argv} &
 		breaksw
 	
@@ -108,11 +105,14 @@ tt		shift
 
 	case 'Media:Social':
 		shift
-		set rtorrent_session_dir = "`/usr/bin/grep -r session ~/.rtorrent.rc | sed 's/session\ =\ \(.*\)/\1'/`"
-		if ( "${rtorrent_session_dir}" != "" && -d "${rtorrent_session_dir}" ) then
-			if ( -e "${rtorrent_session_dir}/rtorrent.lock" ) rm -f "${rtorrent_session_dir}/rtorrent.lock"
-			#if ( -e "${rtorrent_session_dir}/rtorrent.dht_cache" ) rm -f "${rtorrent_session_dir}/rtorrent.dht_cache"
-		end
+		if ( -e "${HOME}/.rtorrent.rc" ) then
+			set rtorrent_session_dir = "`/usr/bin/grep session '${HOME}/.rtorrent.rc' | /usr/bin/sed 's/^[^=]\+=\ \(.*\)$/\1/g'`"
+			if ( "${rtorrent_session_dir}" != "" && -d "${rtorrent_session_dir}" ) then
+				if ( -e "${rtorrent_session_dir}/rtorrent.lock" ) rm -f "${rtorrent_session_dir}/rtorrent.lock"
+				if ( -e "${rtorrent_session_dir}/rtorrent.dht_cache" ) rm -f "${rtorrent_session_dir}/rtorrent.dht_cache"
+			endif
+		endif
+		
 		/usr/bin/gnome-terminal \
 			--hide-menubar \
 			--geometry=${alacast_geometry} \
@@ -121,7 +121,7 @@ tt		shift
 			--tab-with-profile="projects" --title="media" --working-directory="/media" \
 			--tab-with-profile="projects" --title="podiobooks" --working-directory="/media/podiobooks" \
 			--tab-with-profile="projects" --title="podcasts" --working-directory="/media/podcasts" \
-			--tab-with-profile="projects" --title="Alacast" --working-directory="/projects/www/Alacast/version2" \
+			--tab-with-profile="projects" --title="Alacast" --working-directory="/projects/gtk/Alacast" \
 			--tab-with-profile="projects" --title="uberChicGeekChicks-Podcast-Syncronizer" --working-directory="/projects/www/Alacast/bin" --command="/projects/www/Alacast/bin/uberChicGeekChicks-Podcast-Syncronizer.php --update=detailed --logging --player=xine --interactive" \
 		${argv} &
 		breaksw
@@ -144,12 +144,12 @@ tt		shift
 			--tab-with-profile="screen" --title="screen" --command="/usr/bin/screen -${screens_options}"  \
 			--tab-with-profile="projects" --title="media" --working-directory="/media" \
 			--tab-with-profile="projects" --title="/profile.d" --working-directory="/profile.d" \
-			--tab-with-profile="projects" --title="~/" --working-directory="${HOME}" \
 			--tab-with-profile="projects" --title="programs" --working-directory="/programs" \
 			--tab-with-profile="projects" --title="srv" --working-directory="/srv" \
 			--tab-with-profile="projects" --title="ssh" --working-directory="/projects/ssh" \
 			--tab-with-profile="projects" --title="projects" --working-directory="/projects" \
-			--tab-with-profile="projects" --title="uberChicGeekChick.Com" --working-directory="/projects/www/MyWebDesigns/uberChicGeekChick.Com" \
+			--tab-with-profile="projects" --title="uberChicGeekChick.Com" --working-directory="/projects/www/MyWebDesigns/mirrors/uberChicGeekChick.Com" \
+			--tab-with-profile="projects" --title="~/" --working-directory="${HOME}" \
 		${argv} &
 		breaksw
 	endsw
